+++ GUI/LightingGUI/settingspage.cpp
@@ -8,6 +8,7 @@
 #include "ui_settingspage.h"
 
 #include <QDebug>
+#include <QSignalMapper>
 
 SettingsPage::SettingsPage(QWidget *parent) :
     QWidget(parent),
@@ -15,30 +16,71 @@ SettingsPage::SettingsPage(QWidget *parent) :
     ui->setupUi(this);
 
     // setup sliders
+    mSliderSpeedValue = 425;
     ui->speedSlider->slider->setRange(1, 1000);
-    ui->speedSlider->slider->setValue(300);
-    ui->speedSlider->setSliderHeight(0.6f);
+    ui->speedSlider->slider->setValue(mSliderSpeedValue);
+    ui->speedSlider->setSliderHeight(0.5f);
+    ui->speedSlider->slider->setTickPosition(QSlider::TicksBelow);
+    ui->speedSlider->slider->setTickInterval(100);
+
     ui->timeoutSlider->slider->setRange(0,240);
     ui->timeoutSlider->slider->setValue(120);
-    ui->timeoutSlider->setSliderHeight(0.6f);
+    ui->timeoutSlider->setSliderHeight(0.5f);
+    ui->timeoutSlider->slider->setTickPosition(QSlider::TicksBelow);
+    ui->timeoutSlider->slider->setTickInterval(40);
 
-    // setup the signals and slots
+    QSignalMapper *commTypeMapper = new QSignalMapper(this);
+#ifndef MOBILE_BUILD
+    ui->serialButton->setCheckable(true);
+    connect(ui->serialButton, SIGNAL(clicked(bool)), commTypeMapper, SLOT(map()));
+    commTypeMapper->setMapping(ui->serialButton, (int)ECommType::eSerial);
+#else
+    // hide PC-specific elements
+    ui->serialButton->setHidden(true);
+    ui->availablePortsLabel->setHidden(true);
+#endif //MOBILE_BUILD
+
+    ui->httpButton->setCheckable(true);
+    connect(ui->httpButton, SIGNAL(clicked(bool)), commTypeMapper, SLOT(map()));
+    commTypeMapper->setMapping(ui->httpButton, (int)ECommType::eHTTP);
+
+    ui->udpButton->setCheckable(true);
+    connect(ui->udpButton, SIGNAL(clicked(bool)), commTypeMapper, SLOT(map()));
+    commTypeMapper->setMapping(ui->udpButton, (int)ECommType::eUDP);
+
+    connect(commTypeMapper, SIGNAL(mapped(int)), this, SLOT(commTypeSelected(int)));
     connect(ui->speedSlider, SIGNAL(valueChanged(int)), this, SLOT(speedChanged(int)));
     connect(ui->timeoutSlider, SIGNAL(valueChanged(int)), this, SLOT(timeoutChanged(int)));
-
-    connect(ui->serialList, SIGNAL(itemClicked(QListWidgetItem*)), this, SLOT(listClicked(QListWidgetItem*)));
+    connect(ui->connectionList, SIGNAL(itemClicked(QListWidgetItem*)), this, SLOT(listClicked(QListWidgetItem*)));
+    connect(ui->plusButton, SIGNAL(clicked(bool)), this, SLOT(plusButtonClicked()));
+    connect(ui->minusButton, SIGNAL(clicked(bool)), this, SLOT(minusButtonClicked()));
 }
 
 SettingsPage::~SettingsPage() {
     delete ui;
 }
 
+
 // ----------------------------
 // Slots
 // ----------------------------
 
 void SettingsPage::speedChanged(int newSpeed) {
-    mData->speed(newSpeed);
+    mSliderSpeedValue = newSpeed;
+    int finalSpeed;
+    // first half of slider is going linearly between 20 FPS down to 1 FPS
+    if (newSpeed < 500) {
+        float percent = newSpeed / 500.0f;
+        finalSpeed = (int)((1.0f - percent) * 2000.0f);
+    } else {
+        // second half maps 1FPS to 0.01FPS
+        float percent = newSpeed - 500.0f;
+        finalSpeed = (500 - percent) / 5.0f;
+        if (finalSpeed < 2.0f) {
+            finalSpeed = 2.0f;
+        }
+    }
+    mData->speed((int)finalSpeed);
     mComm->sendSpeed(mData->speed());
 }
 
@@ -49,21 +91,91 @@ void SettingsPage::timeoutChanged(int newTimeout) {
 
 //TODO:COMM LISTS
 void SettingsPage::listClicked(QListWidgetItem* item) {
-    // only change connection if a new one is clicked
-    if (QString::compare(item->text(), mComm->comm()->name())) {
-        // check a mac os x case where it adds a .cu for certain connections
-        QString cuCheck = QString("cu.%1").arg(item->text());
-        if (!QString::compare(cuCheck, mComm->comm()->name())) {
-            // do nothing
-        } else {
-            // close the current connection
-            //mComm->comm()->closeConnection();
-            // connect to the selected connection
-           // mComm->mSerial->connectSerialPort(item->text());
-        }
+    //qDebug() << item->text() << "vs" << mCurrentListString;
+    if (QString::compare(item->text(), mCurrentListString)) {
+        mCurrentListString = item->text();
+#ifndef MOBILE_BUILD
+        // serial connections require a little more effort to properly
+        // switch, they need to be closed first and then reconnected
+        mComm->closeCurrentConnection();
+        mComm->changeSerialPort(mCurrentListString);
+#endif //MOBILE_BUILD
+        mComm->comm()->selectConnection(mCurrentListString);
+        ui->lineEdit->setText(mComm->comm()->currentConnection());
     }
 }
 
+
+void SettingsPage::highlightButton(ECommType currentCommType) {
+    ui->serialButton->setChecked(false);
+    ui->httpButton->setChecked(false);
+    ui->udpButton->setChecked(false);
+
+    if (mComm->currentCommType() == ECommType::eHTTP) {
+        ui->httpButton->setChecked(true);
+    } else if (mComm->currentCommType() == ECommType::eUDP) {
+        ui->udpButton->setChecked(true);
+    }
+#ifndef MOBILE_BUILD
+    else if (mComm->currentCommType() == ECommType::eSerial) {
+        ui->serialButton->setChecked(true);
+    }
+#endif //MOBILE_BUILD
+}
+
+
+void SettingsPage::commTypeSelected(int type) {
+    if ((ECommType)type != mComm->currentCommType()) {
+        mComm->currentCommType((ECommType)type);
+        highlightButton((ECommType)type);
+        updateConnectionList();
+        ui->connectionList->item(mComm->comm()->selectedIndex())->setSelected(true);
+    }
+#ifndef MOBILE_BUILD
+        if ((ECommType)type == ECommType::eSerial) {
+            ui->lineEdit->setHidden(true);
+            ui->plusButton->setHidden(true);
+            ui->minusButton->setHidden(true);
+            ui->availablePortsLabel->setHidden(false);
+        } else {
+            ui->lineEdit->setHidden(false);
+            ui->plusButton->setHidden(false);
+            ui->minusButton->setHidden(false);
+            ui->availablePortsLabel->setHidden(true);
+        }
+#endif //MOBILE_BUILD
+
+}
+
+
+void SettingsPage::plusButtonClicked() {
+    bool isSuccessful = mComm->comm()->addConnection(ui->lineEdit->text());
+    if (isSuccessful) {
+        // adjusts the backend to the connection
+        mComm->comm()->selectConnection(ui->lineEdit->text());
+        // updates the connection list
+        updateConnectionList();
+        // selects the top connection on the GUI level
+        ui->connectionList->item(0)->setSelected(true);
+    }
+}
+
+
+void SettingsPage::minusButtonClicked() {
+    bool isSuccessful = mComm->comm()->removeConnection(mCurrentListString);
+    if (isSuccessful) {
+        mCurrentListString = ui->connectionList->item(0)->text();
+        // adjusts the backend to the connection
+        mComm->comm()->selectConnection(mCurrentListString);
+        // update the line edit text
+        ui->lineEdit->setText(mCurrentListString);
+        // updates the connection list
+        updateConnectionList();
+        // selects the top connection on the GUI level
+        ui->connectionList->item(0)->setSelected(true);
+    }
+
+}
 
 // ----------------------------
 // Protected
@@ -72,10 +184,10 @@ void SettingsPage::listClicked(QListWidgetItem* item) {
 
 void SettingsPage::showEvent(QShowEvent *event) {
     Q_UNUSED(event);
-    // set the sliders to their proper values
-    ui->speedSlider->slider->setValue(mData->speed());
+    ui->speedSlider->slider->setValue(mSliderSpeedValue);
     ui->timeoutSlider->slider->setValue(mData->timeOut());
-
+    highlightButton(mComm->currentCommType());
+    commTypeSelected((int)mComm->currentCommType());
     // default the settings bars to the current colors
     if (mData->currentRoutine() <= ELightingRoutine::eSingleGlimmer) {
         ui->speedSlider->setSliderColorBackground(mData->mainColor());
@@ -84,19 +196,17 @@ void SettingsPage::showEvent(QShowEvent *event) {
         ui->speedSlider->setSliderColorBackground(mData->colorsAverage(mData->currentColorGroup()));
         ui->timeoutSlider->setSliderColorBackground(mData->colorsAverage(mData->currentColorGroup()));
     }
-#ifndef MOBILE_BUILD
 
     // update the serial list, if needed
-    updateSerialList();
-
-    // select the QSerialPort portname that matches the currently connected port
-    for (int i = 0; i < ui->serialList->count(); i++) {
-        QListWidgetItem *item = ui->serialList->item(i);
-        if (!QString::compare(item->text(), mComm->comm()->name())) {
-            item->setSelected(true);
-        } else {
-            item->setSelected(false);
-        }
+    updateConnectionList();
+#ifndef MOBILE_BUILD
+    if (mComm->currentCommType() != ECommType::eSerial) {
+#endif //MOBILE_BUILD
+      ui->connectionList->item(mComm->comm()->selectedIndex())->setSelected(true);
+      mCurrentListString = ui->connectionList->item(0)->text();
+#ifndef MOBILE_BUILD
+    } else {
+        ui->connectionList->item(mComm->comm()->selectedIndex())->setSelected(true);
     }
 #endif //MOBILE_BUILD
 }
@@ -106,21 +216,20 @@ void SettingsPage::showEvent(QShowEvent *event) {
 // Private
 // ----------------------------
 
-//TODO:COMM LISTS
-void SettingsPage::updateSerialList() {
-//    // look for updates in the serial port list
-//    for (QSerialPortInfo port : mComm->mSerial->serialList) {
-//        bool itemFound = false;
-
-//        for (int i = 0; i < ui->serialList->count(); i++) {
-//            QListWidgetItem *item = ui->serialList->item(i);
-//            if (!QString::compare(item->text(), port.portName())) {
-//                itemFound = true;
-//            }
-
-//        }
-//        if (!itemFound) {
-//            ui->serialList->addItem(port.portName());
-//        }
-//    }
+void SettingsPage::updateConnectionList() {
+    ui->connectionList->clear();
+    bool firstItem = true;
+    for (QString connectionName : (*mComm->comm()->connectionList().get())) {
+        bool itemFound = false;
+        for (int i = 0; i < ui->connectionList->count(); i++) {
+            QListWidgetItem *item = ui->connectionList->item(i);
+            if (!QString::compare(item->text(), connectionName)) {
+                itemFound = true;
+            }
+        }
+        if (!itemFound) {
+            ui->connectionList->addItem(connectionName);
+        }
+    }
+    ui->lineEdit->setText(mComm->comm()->currentConnection());
 }